Wallet Analytics Form setup complete
