+++ my_app/src/app/components/TransactionBalanceInfoTable.tsx
@@ -1,0 +1,6 @@
+// Transaction Balance Info Table Custom Component
+export default function TransactionBalanceInfoTable() {
+    return (
+        <div></div>
+    )
+}
