+++ my_app/src/app/wallet-analytics/page.tsx
@@ -1,24 +1,5 @@
-'use client'
-
-import { useState } from 'react';
-import { useRouter } from 'next/navigation';
-import { Input } from "../components/ui/input";
-import { Button } from "../components/ui/button";
-import { Card, CardContent, CardHeader, CardTitle, CardDescription } from "../components/ui/card";
-import { Alert, AlertDescription, AlertTitle } from "../components/ui/alert";
-import NetworkSelector from '../components/NetworkSelector';
-
 // Wallet Analytics Custom Component
 export default function WalletAnalyticsPage() {
-    const router = useRouter()
-    const [walletAddress, setWalletAddress] = useState("")
-    const [network, setNetwork] = useState("eth")
-    const [showAlert, setShowAlert] = useState(false)
-
-    const handleSubmit = (e: React.FormEvent) => {
-        e.preventDefault()
-        // Handle form submission logic here
-    }
     
     // Render the Wallet Analytics Page Component
     return (
@@ -32,42 +13,6 @@ export default function WalletAnalyticsPage() {
                 <p className="text-xl text-gray-400 mb-12 text-center">
                     Dive deep into wallet performance and analytics
                 </p>
-                {showAlert && (
-                    <Alert variant="destructive" className="mb-6">
-                        <AlertTitle>Error</AlertTitle>
-                        <AlertDescription>
-                            There was an error processing your request. Please try again.
-                        </AlertDescription>
-                    </Alert>
-                )}
-                <Card className="bg-gray-900 border-gray-800 shadow-xl w-full">
-                    <CardHeader className="border-b border-gray-800 pb-6">
-                        <CardTitle className="text-3xl font-bold text-gray-100">Analyze Wallet</CardTitle>
-                        <CardDescription className="text-gray-400 text-lg font-light">
-                            Enter wallet details for in-depth analysis
-                        </CardDescription>
-                    </CardHeader>
-                    <CardContent className="space-y-6 pt-6">
-                        <form onSubmit={handleSubmit} className="space-y-4">
-                            <Input
-                                placeholder="Enter Wallet Address"
-                                value={walletAddress}
-                                onChange={(e) => setWalletAddress(e.target.value)}
-                                className="w-full bg-gray-800 text-gray-100 border-gray-700 focus:ring-gray-400 placeholder-gray-500"
-                                required
-                            />
-                            <NetworkSelector />
-                            <div className="flex justify-center space-x-4 pt-4">
-                                <Button 
-                                    type="submit"
-                                    className="bg-gradient-to-r from-gray-600 to-gray-400 text-white py-2 px-6 rounded-md hover:from-gray-500 hover:to-gray-300 transition-all duration-300 transform hover:scale-105 font-medium"
-                                >
-                                    Analyze
-                                </Button>
-                            </div>
-                        </form>
-                    </CardContent>
-                </Card>
             </div>
         </div>
     )
